Fix Break/Repair/Enqueue buttons by switching IDs to uint32_t
Symptom (in webtest deploy):
- + Voltz Classic / Hyper Bolt / Aurora Zero buttons appeared to do nothing.
- Break / Repair buttons appeared to do nothing — HP stayed at 100.

Root cause:
ctrl::LineId and ctrl::MachineId were std::uint64_t. Emscripten's embind
does not transparently marshal uint64_t to/from JS Number without the
-sWASM_BIGINT flag, which we don't (yet) enable. Snapshot data still
displayed because reads happened to coerce, but the IDs JS passed back
into controller.enqueue / .breakMachine / .repair were being mangled, so
the BE looked them up as nonsense IDs and returned NotFound.

Fix:
Narrow ctrl::LineId and ctrl::MachineId to std::uint32_t. Cast at the
gactorio::* boundary inside Controller.cpp. We have a single line and
four machines, so the narrowing is lossless.

Also rebuild docs/gactorio.{js,wasm} from the updated sources so the
GitHub Pages deploy picks up the fix on the webtest branch.

+++ include/controller/Controller.h
@@ -19,8 +19,9 @@
 
 namespace ctrl {
 
-using LineId    = std::uint64_t;
-using MachineId = std::uint64_t;
+// uint32 keeps embind happy without -sWASM_BIGINT and is plenty for our IDs.
+using LineId    = std::uint32_t;
+using MachineId = std::uint32_t;
 
 enum class ProductKind {
     Unknown,
